+++ src/steps/pytorch/callbacks.py
@@ -4,7 +4,7 @@ from datetime import datetime, timedelta
 from deepsense import neptune
 from torch.optim.lr_scheduler import ExponentialLR
 
-from steps.utils import get_logger
+from ...steps.utils import get_logger
 from .utils import Averager, save_model
 from .validation import score_model
